Added support for tracking a user and also changed the names of some of the packages and class names.
