+++ Documentation/KinectController/com/lcsc/hackathon/kinectcontroller/EsperHandler.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.lcsc.hackathon.kinectcontroller.EsperHandler</li>
+<li>com.lcsc.hackathon.kinectcontroller.esper.EsperHandler</li>
 </ul>
 </li>
 </ul>
