+++ Documentation/KinectController/com/lcsc/hackathon/kinectcontroller/KinectHandler.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.lcsc.hackathon.kinectcontroller.kinect.KinectHandler</li>
+<li>com.lcsc.hackathon.kinectcontroller.kinect.KinectUserTracker</li>
 </ul>
 </li>
 </ul>
