+++ Documentation/KinectController/overview-summary.html
@@ -104,7 +104,7 @@
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a href="com/lcsc/hackathon/kinectcontroller/esperlisteners/package-summary.html">com.lcsc.hackathon.kinectcontroller.esperlisteners</a></td>
+<td class="colFirst"><a href="com/lcsc/hackathon/kinectcontroller/esperlisteners/package-summary.html">com.lcsc.hackathon.kinectcontroller.esper</a></td>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
