+++ Documentation/KinectController/package-list
@@ -4,5 +4,5 @@ com.lcsc.hackathon.kinectcontroller.controller
 com.lcsc.hackathon.kinectcontroller.emulation
 com.lcsc.hackathon.kinectcontroller.emulation.reactions
 com.lcsc.hackathon.kinectcontroller.emulation.reactions.config
-com.lcsc.hackathon.kinectcontroller.esperlisteners
+com.lcsc.hackathon.kinectcontroller.esper
 com.lcsc.hackathon.kinectcontroller.posturerules
