+++ src/main/java/com/lcsc/hackathon/kinectcontroller/esper/EventListener.java
@@ -23,7 +23,7 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 
-package com.lcsc.hackathon.kinectcontroller.esperlisteners;
+package com.lcsc.hackathon.kinectcontroller.esper;
 
 import com.espertech.esper.client.UpdateListener;
 import com.espertech.esper.client.EventBean;
