+++ src/main/java/com/lcsc/hackathon/kinectcontroller/kinect/KinectDebugWindow.java
@@ -208,7 +208,6 @@ public class KinectDebugWindow extends Component implements UserTracker.NewFrame
             }
 
             depthFrame.release();
-            depthFrame = null;
         }
 
         repaint();
