+++ src/main/java/com/lcsc/hackathon/kinectcontroller/kinect/KinectUserTracker.java
@@ -26,6 +26,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 package com.lcsc.hackathon.kinectcontroller.kinect;
 
 import com.lcsc.hackathon.kinectcontroller.controller.ControllerStateMachine;
+import com.lcsc.hackathon.kinectcontroller.posturerules.Angle;
 import com.lcsc.hackathon.kinectcontroller.posturerules.Rule;
 import com.primesense.nite.*;
 import org.openni.Device;
@@ -43,13 +44,14 @@ import java.util.List;
  * This class will be tasked with keeping track of the libraries that are needed to talk to the Kinect. So this class will
  * be handing off the positions of joints to Esper and Esper will then use that data to detect gestures.
  */
-public class KinectHandler implements UserTracker.NewFrameListener{
-    private static final Logger                 _logger = LoggerFactory.getLogger(KinectHandler.class);
+public class KinectUserTracker implements UserTracker.NewFrameListener{
+    private static final Logger                 _logger = LoggerFactory.getLogger(KinectUserTracker.class);
     public         final KinectDebugWindow      kinectWindow;
     private              UserTracker            _tracker;
     private        final ControllerStateMachine _csm;
+    private              short                  _userId = -1;
 
-    public KinectHandler(ControllerStateMachine csm, boolean debug) {
+    public KinectUserTracker(ControllerStateMachine csm, boolean debug) {
         _csm = csm;
 
         System.out.println("Initializing the things.");
@@ -81,9 +83,51 @@ public class KinectHandler implements UserTracker.NewFrameListener{
      *                there is already a copy of it saved in the class' private scope.
      */
     public void onNewFrame(UserTracker tracker) {
-        //TODO Pull the posturerules Event Beans from the current ControllerState, fill those beans with data and give them to Esper.
-        for (Rule rule : _csm.getCurrentRules()) {
-            //Update that rule!
+        UserTrackerFrameRef frame   = tracker.readFrame();
+        UserData userData           = null;
+
+        //These conditions below make it so that there is always only 1 user being tracked and we won't do
+        //anything if there isn't a user being tracked.
+        if (_userId == -1) {
+            List<UserData> users = frame.getUsers();
+            if (users.size() == 0) {
+                return;
+            }
+            else {
+                for (UserData user : users) {
+                    if (user.isNew()) {
+                        _userId = user.getId();
+                        _tracker.startSkeletonTracking(_userId);
+                        userData = user;
+                    }
+                }
+            }
+        }
+        else {
+            userData = frame.getUserById(_userId);
+            if (userData.isLost()) {
+                _tracker.stopSkeletonTracking(_userId);
+                _userId = -1;
+                return;
+            }
+        }
+        if (userData != null) {
+            Skeleton skeleton = userData.getSkeleton();
+            if (skeleton.getState() == SkeletonState.TRACKED) {
+                //Here we will get the posturerule event beans from the current ControllerState.
+                //We'll then update them with data from the user that's being tracked and we'll pass
+                //those objects over to Esper for processing.
+                for (Rule rule : _csm.getCurrentRules()) {
+                    switch (rule.getType()) {
+                        case ANGLE:
+                            Angle angRule           = (Angle) rule;
+                            SkeletonJoint end1      = skeleton.getJoint(JointType.fromNative(angRule.getEnd1()));
+                            SkeletonJoint vertex    = skeleton.getJoint(JointType.fromNative(angRule.getVertex()));
+                            SkeletonJoint end2      = skeleton.getJoint(JointType.fromNative(angRule.getEnd2()));
+                            break;
+                    }
+                }
+            }
         }
     }
 }
